+++ server/CHANGELOG.md
@@ -15,7 +15,27 @@ not generally be shown to end users.
 
 # Changelog
 
-# CURRENT
+# 2.1.16
+
+- [BUG] EXCEED GEAR scores can now be imported for real for real.
+
+# 2.1.15
+
+- [BUG] EXCEED GEAR scores can now be imported for real.
+
+# 2.1.14
+
+- [BUG] Fixed bug where folders/tables wouldn't sync automatically from the seeds, again.
+
+# 2.1.13
+
+- [BUG] Fixed bug where folders/tables wouldn't sync automatically from the seeds.
+
+# 2.1.12
+
+- [ENH] EXCEED GEAR scores can now be imported.
+
+# 2.1.11
 
 - [ENH] The LR2Hook now saves courses.
 - [ENH] The LR2Hook now saves gauge information.